Add infra files
The `instance` file is the JSON for creating the gcloud instance.
The `run-seed.sh` file is the entry point for generating the key and
starting the seed node.
The setup script will be used by the gcloud instance for setting itself
up.
The seed.service file is the systemd file we'll kick off for running the
seed node.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>