[seed] Allow CORS requests for certain routes
This makes the /projects/{urn} and /info paths available
for requesting from other websites, allowing for external UIs
to use seeds as data sources for project metadata as well as
querying the public state of a seed.

Note that these routes were already open/public, this simply
allows them to be queried from a browser as well.