seed: Pin link dependency
As linkd is the main binary going forward to provide seed deployments
the crate that was pulled in from the link repository so far was
sunsetted is gone. Therefore for now link is pinned, which implies no
future upgrades shall be pulled in.

The radicle-bins repository is headed towards the archive.

Signed-off-by: xla <self@xla.is>